feature/sharded-cache: Sharded the cache
Instead of a single map[string]interface{} in which all of
the data is written and against which all threads will have
to get a lock, the cache is now sharded.

Refactored the various structs for dependency injection.

Got rid of the build flags/tags to separate unit from
integration tests.  Added the ability to skip the
integration tests by omitting the export of an env var.
 Kind of a moot point, since all there is right now is
integration testing, but it was mostly to get the model
sorted out for other projects.